Changed models

+++ backend/zerobite/product/models.py
@@ -21,7 +21,7 @@ class Product(models.Model):
   category = models.ForeignKey(Category, on_delete= models.CASCADE, related_name = "products")
   name = models.CharField(max_length=50)
   description = models.TextField()
-  price = models.DecimalField(max_digits = 8, decimal_places = 2)
+  price = models.FloatField()
   image = models.ImageField(upload_to='products/', blank = True, null = True)
   manufactured_date = models.DateField()
   expiry_date = models.DateField()
